+++ munchkin/munchkin.py
@@ -37,12 +37,6 @@ def password_dumper(args, card):
 
 # Generate card similar to ones on http://passwordcard.org
 def generate_card_pcard(args, card):
-    # FIXME: Symbols feature is broken
-    if args.symbols:
-        logging.error("Sorry, this feature is broken at the moment. :(")
-        logging.error("Please select the custom card type and paste your card manually.")
-        sys.exit(3)
-
     card.generate_password_card(symbols=args.symbols, digits=args.digits)
     password_dumper(args, card)
 
@@ -79,6 +73,10 @@ def menu():
     wordlist_options.add_argument('-r', '--right-to-left', help='read card from bottom right to top left', action='store_true')
     wordlist_options.add_argument('-t', '--top-down', help='read card from top left to bottom left', action='store_true')
     wordlist_options.add_argument('-b', '--bottom-up', help='read card from bottom right to top right', action='store_true')
+    wordlist_options.add_argument('-z', '--zig-zag', help='read card zig-zag from top left', action='store_true')
+    wordlist_options.add_argument('-x', '--zig-zag-rev', help='read card zig-zag from bottom right', action='store_true')
+    wordlist_options.add_argument('-d', '--diagonal', help='read card diagonally', action='store_true')
+    wordlist_options.add_argument('-a', '--all', help='use each read strategy from above', action='store_true')
 
     subparsers = parser.add_subparsers(title='subcommands', description='valid subcommands', help='Generate wordlist based on cards from:')
     parser_pc = subparsers.add_parser('pcard', help="http://passwordcard.org", parents=[parent_parser])
@@ -86,7 +84,7 @@ def menu():
 
     passwordcard_options = parser_pc.add_argument_group('passwordcard options')
     passwordcard_options.add_argument('-s', '--seed', help='card number (e.g. 7eb3fbfa560d1d1e)', required=True, metavar='str')
-    passwordcard_options.add_argument('--symbols', help='include symbols (broken)', action='store_true')
+    passwordcard_options.add_argument('--symbols', help='include symbols', action='store_true')
     passwordcard_options.add_argument('--digits', help='incude digits', action='store_true')
 
     # Set controller functions
@@ -99,8 +97,15 @@ def menu():
 
 def controller(args):
     # Verifies whether at least one strategy was selected
-    if (args.left_to_right or args.right_to_left or args.top_down or args.bottom_up) is False:
-        logging.error("Please select at least one algorithm (hint: -l/-r/-t/-b)\r\n")
+    if (args.left_to_right or
+        args.right_to_left or
+        args.top_down or
+        args.bottom_up or
+        args.zig_zag or
+        args.zig_zag_rev or
+        args.diagonal or
+        args.all) is False:
+        logging.error("Please select at least one algorithm (hint: -l/-r/-t/-b/-z/-x/-d/--all)\r\n")
         sys.exit(3)
     else:
         # Create card and start processing